refactor: move hooks and UI shell to core/
