+++ src/core/ui/CameraHints.css
