+++ src/core/ui/CameraHints.tsx
