+++ src/core/ui/Sidebar.tsx
@@ -1,8 +1,8 @@
-// src/components/Sidebar.tsx
+// src/core/ui/Sidebar.tsx
 import './Sidebar.css'
-import piecesConfig from '../data/pieces-config.json'
-import type { PiecesConfig, PieceCategory } from '../core/types'
-import { useStore } from '../core/store/useStore'
+import piecesConfig from '../../data/pieces-config.json'
+import type { PiecesConfig, PieceCategory } from '../types'
+import { useStore } from '../store/useStore'
 
 const config = piecesConfig as PiecesConfig
